mkimg: Compress images that are too long
Based on the size of the clusters, qcow2 format sets a limit on the
offset where a compressed cluster can be referenced. For example, if the
limit is "x" and we compress a cluster, then its host offset cannot be
greater than x because its cluster descriptor cannot "point" there. In
that case, we will not compress the cluster.