mkimg: Allocate only one cluster for rc blocks
mkimg had been allocating a set of continous memory in the heap which
held the reference count blocks. When all the reference count blocks were
filled they would be written into the file. However, for lengthy images
this memory could be of significant size. This patch modifies the
qcow_rcblks function and performs multiple writes instead of a single
one.